Extract getTargetAmount into shared module
The embedded calculator and the authoritative-sources rebuild were
carrying byte-for-byte identical per-record classification logic. A
future schema change to one would silently diverge the embedded balance
the calculator records from the totals the rebuild reproduces — exactly
the divergence the diagnostic exists to surface.
